Bumped to 1.3. Fixed file handling for prboom+
Apparently prboom+ handles -file in a way no other port does -
it expects -file to be followed by all the files and passing each
file with -file doesn't seem to work. All files that would be added with -file
are now sorted to the end and added all at once to the end of the
command line with one instance of -file. This seems to work for everything.
I moved the additional params to the beginning to minimize possible interference.